+++ Userland/Libraries/LibGfx/CMakeLists.txt
@@ -33,6 +33,7 @@ set(SOURCES
     ICC/WellKnownProfiles.cpp
     ImageFormats/BMPLoader.cpp
     ImageFormats/BMPWriter.cpp
+    ImageFormats/BooleanDecoder.cpp
     ImageFormats/DDSLoader.cpp
     ImageFormats/GIFLoader.cpp
     ImageFormats/ICOLoader.cpp
